Updated sprites backend

+++ mapview/sprites.h
@@ -45,7 +45,7 @@ void set_projection(int x, int y, int w, int h);
 
 
 bool init_things(void);
-sprite_t* get_thing_sprite_name(int thing_type, int angle);
+sprite_t* get_thing_sprite_name(uint16_t thing_type, uint16_t angle);
 void cleanup_things(void);
 
 extern int screen_width;
